Fix for almost straight lines
If the 3 points are almost on a straight line and the middle one is moved slightly the radius calculated in some situations becomes small. Adding this makes points that lie only 0.1m offset from one another to be perfectly straight.
Open street map points have an accuracy of 1cm and an average road is about 3m wide. 
Having a lot of points on a road close together will cause such problems. Maybe a future version could check the distance between points and eliminate 2 by interpolating inbetween until there is a minimum distance of 6m about double the road width.